SAGE-1622: fix PxE boot failure of new RPi 4 hardware
The new RPi 4 hardware bootloader requires the latest start4.elf
(and other files) in the tftp PxE boot folder. Therefore we are cloning
a more recent firmware (which was released after the listed date in the
RPi bootloader - early 2022) and syncing that on top of the 20.04.2 image.

This was done in favor of upgrading to the latest 20.04 (20.04.4) image
as 20.04.4 comes with a new kernel (i.e. vmlinux) which would require
an entire OS update.